:boom: Drop short switches for SSH opts
There are too many SSH config options alltogether to keep sensible,
short config mappings for them. They're not readable in the CLI and
can conflict with other switches, that are related to the general
processing flow.

That being said, short switches in other categories are still
allowed (e.g. `-o` for output file) for the time being.

This is a breaking change, but since the script was not made public
yet, it's not going to impact the *major* version at all.